// WADNR.Web/src/app/shared/generated/model/find-your-forester-question-tree-node.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class FindYourForesterQuestionTreeNode { 
    FindYourForesterQuestionID?: number;
    QuestionText?: string | null;
    ForesterRoleID?: number | null;
    ForesterRoleDisplayName?: string | null;
    ForesterRoleName?: string | null;
    ResultsBonusContent?: string | null;
    Children?: Array<FindYourForesterQuestionTreeNode> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface FindYourForesterQuestionTreeNodeForm { 
    FindYourForesterQuestionID?: FormControl<number>;
    QuestionText?: FormControl<string>;
    ForesterRoleID?: FormControl<number>;
    ForesterRoleDisplayName?: FormControl<string>;
    ForesterRoleName?: FormControl<string>;
    ResultsBonusContent?: FormControl<string>;
    Children?: FormControl<Array<FindYourForesterQuestionTreeNode>>;
}

export class FindYourForesterQuestionTreeNodeFormControls { 
    public static FindYourForesterQuestionID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static QuestionText = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ResultsBonusContent = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Children = (value: FormControlState<Array<FindYourForesterQuestionTreeNode>> | Array<FindYourForesterQuestionTreeNode> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<FindYourForesterQuestionTreeNode>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}