// WADNR.Web/src/app/shared/generated/model/forester-role-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ForesterRoleLookupItem { 
    ForesterRoleID?: number;
    ForesterRoleDisplayName?: string | null;
    ForesterRoleName?: string | null;
    SortOrder?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ForesterRoleLookupItemForm { 
    ForesterRoleID?: FormControl<number>;
    ForesterRoleDisplayName?: FormControl<string>;
    ForesterRoleName?: FormControl<string>;
    SortOrder?: FormControl<number>;
}

export class ForesterRoleLookupItemFormControls { 
    public static ForesterRoleID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ForesterRoleName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static SortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}